+++ src/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 /**
  * # Mobius
  *
@@ -5,12 +6,9 @@
  * ! Dark art ahead
  *
  * ? Total hours wasted by this: 0
- * ? Please send PR to update ^
  *
  * @author saltyAom
  */
-
-type Defer<T> = { __defer: T }
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
 type Whitespace =
@@ -67,19 +65,24 @@ type ExtractType<T extends string> = T extends `${infer Type}\n${infer Rest}`
 type CustomTypes = Record<string, string | Record<string, unknown>>
 type Scalar = Record<string, unknown>
 
+/**
+ * Actually a string
+ */
+type ID = string
+
 type GQLTypes = {
     String: string
     Int: number
     Float: number
     Boolean: boolean
-    ID: string
+    ID: ID
 }
 
 type MergeInterface<
     Interfaces extends string,
     Known extends CustomTypes = {},
     Types extends Record<string, unknown> = {}
-> = Interfaces extends `${infer Name},${infer Rest}`
+> = Interfaces extends `${infer Name}${',' | '&'}${infer Rest}`
     ? MergeInterface<
           Rest,
           Known,
@@ -107,57 +110,27 @@ type RemoveMultiLineComment<T extends string> =
         ? `${First}${RemoveMultiLineComment<Rest>}`
         : T
 
-// Name extends keyof T
-//     ? K extends {
-//           __defer: infer Eager extends [Defer<string>]
-//       }
-//         ? T & {
-//               __defer: Eager
-//           }
-//         : T
-//     : T
-
-type CreateInnerMobius<
+export type CreateInnerMobius<
     T extends string,
-    Scalars extends Scalar = {},
     Known extends CustomTypes = {}
 > = T extends `${infer Ops}{${infer Schema}}${infer Rest}`
     ? Trim<RemoveComment<Ops>> extends `${infer Keyword} ${infer Name}`
         ? CreateInnerMobius<
               Rest,
-              Scalars,
               Known &
                   (Keyword extends 'type'
                       ? {
                             [name in TrimLeft<FirstWord<Name>>]: Prettify<
-                                MapSchema<
-                                    name,
-                                    RemoveComment<Schema>,
-                                    Known & Scalars
-                                > &
+                                MapSchema<RemoveComment<Schema>, Known> &
                                     (Name extends `${infer _} implements ${infer Interfaces}`
                                         ? MergeInterface<Interfaces, Known>
                                         : {})
                             >
-                        } extends infer A
-                          ? Name extends keyof A
-                              ? A[Name] extends {
-                                    [k in `__defer${string}`]: infer Eager extends string
-                                }
-                                  ? A & {
-                                        __defer: Eager
-                                    }
-                                  : A
-                              : A
-                          : never
+                        }
                       : Keyword extends 'input' | 'interface'
                       ? {
                             [name in TrimLeft<Name>]: Prettify<
-                                MapSchema<
-                                    name,
-                                    RemoveComment<Schema>,
-                                    Known & Scalars
-                                >
+                                MapSchema<RemoveComment<Schema>>
                             >
                         }
                       : Keyword extends 'enum'
@@ -172,19 +145,15 @@ type CreateInnerMobius<
                             Fragment: {
                                 [name in TrimLeft<FirstWord<Name>>]: Prettify<
                                     Name extends `${infer _} on ${infer Target}`
-                                        ? Target extends keyof Known
-                                            ? Pick<
-                                                  Known[Target],
-                                                  NonNullable<
-                                                      Exclude<
-                                                          MapEnum<
-                                                              RemoveComment<Schema>
-                                                          >,
-                                                          ''
-                                                      >
-                                                  >
+                                        ? {
+                                              Target: Target
+                                              Value: Exclude<
+                                                  MapEnum<
+                                                      RemoveComment<Schema>
+                                                  >,
+                                                  '' | null | undefined
                                               >
-                                            : {}
+                                          }
                                         : {}
                                 >
                             }
@@ -195,7 +164,6 @@ type CreateInnerMobius<
                             `${TrimLeft<
                                 GetLastLine<Ops>
                             >}{${RemoveComment<Schema>}}`,
-                            Scalars,
                             Known
                         >
                       : Keyword extends 'union'
@@ -204,8 +172,7 @@ type CreateInnerMobius<
                             `${TrimLeft<
                                 GetLastLine<Ops>
                             >}{${RemoveComment<Schema>}}`,
-                            Scalars,
-                            Known & MapUnion<Ops, Scalars & Known>
+                            Known & MapUnion<Ops, Known>
                         >
                       : /**
                        * ? TypeScript is greedy, scalar can eat other word until
@@ -215,11 +182,7 @@ type CreateInnerMobius<
                       ? TrimLeft<
                             RemoveComment<Ops>
                         > extends `${infer _}\n${infer Prefix}`
-                          ? CreateInnerMobius<
-                                `${Prefix}{${Schema}}`,
-                                Scalars,
-                                Known
-                            >
+                          ? CreateInnerMobius<`${Prefix}{${Schema}}`, Known>
                           : {}
                       : Known)
           >
@@ -263,7 +226,6 @@ type MapEnum<
     : Carry
 
 type MapSchema<
-    SchemaName extends string,
     T extends string,
     Known extends CustomTypes = {}
 > = T extends `${infer Name}:${infer Type}`
@@ -273,129 +235,122 @@ type MapSchema<
                   infer Type extends string,
                   infer Rest extends string
               ]
-                ? (MapType<
-                      `${SchemaName}.${TrimLeft<Name>}`,
-                      Type,
-                      Known
-                  > extends infer Typed
-                      ? {
-                            [word in Name as TrimLeft<Name> extends infer Candidate extends string
-                                ? Candidate extends `#${infer _}`
-                                    ? never
-                                    : Candidate
-                                : never]: Prettify<
-                                MapArgument<SchemaName, Params, Known>
-                            > extends infer Argument
-                                ? Partial<Argument> extends Argument
-                                    ? (p?: Argument) => Typed
-                                    : (p: Argument) => Typed
-                                : never
-                        } & (NonNullable<Typed> extends Defer<
-                            infer A extends string
-                        >
-                            ? {
-                                  [K in `__defer${Name}`]: A
-                              }
-                            : {})
-                      : {}) &
-                      MapSchema<SchemaName, Rest, Known>
+                ? {
+                      [word in Name as TrimLeft<Name> extends infer Candidate extends string
+                          ? Candidate extends `#${infer _}`
+                              ? never
+                              : Candidate
+                          : never]: Prettify<
+                          MapArgument<Params, Known>
+                      > extends infer Argument
+                          ? Partial<Argument> extends Argument
+                              ? (p?: Argument) => FormatType<Type>
+                              : (p: Argument) => FormatType<Type>
+                          : never
+                  } & MapSchema<Rest, Known>
                 : {}
             : {}
         : ExtractType<Type> extends [
               infer Type extends string,
               infer Rest extends string
           ]
-        ? (MapType<`${SchemaName}.${TrimLeft<Name>}`, Type, Known> extends infer Typed
-              ? {
-                    [word in Name as TrimLeft<Name>]: Typed
-                } & (NonNullable<Typed> extends Defer<infer A extends string>
-                    ? {
-                          // ? Non Collision container
-                          [K in `__defer${Name}`]: A
-                      }
-                    : {})
-              : {}) &
-              MapSchema<SchemaName, Rest, Known>
+        ? {
+              [word in Name as TrimLeft<Name>]: FormatType<Type>
+          } & MapSchema<Rest, Known>
         : {}
     : {}
 
-type MapType<
-    SchemaName extends string,
-    T extends string,
-    Known extends CustomTypes = {}
-> = MapInnerType<
-    SchemaName,
-    FirstWord<T> extends `${infer Type}!` ? Type : T,
-    Known
-> extends infer Type
-    ? unknown extends Type
-        ? unknown
-        : T extends `${infer _}!${infer _}`
-        ? Type
-        : Type | null
-    : unknown
+type RemovePrefixArrayBracket<T extends string> = T extends `[${infer Rest}`
+    ? RemovePrefixArrayBracket<Rest>
+    : T
 
-type MapInnerType<
-    SchemaName extends string,
+// @ts-ignore To hard to explain this shape in TS, I works trust me
+type CreateArray<
     T extends string,
-    Known extends CustomTypes = {}
-> = GQLTypes & Known extends infer Types
-    ? T extends keyof Types
-        ? Types[T]
-        : T extends `[${infer InnerType}!]`
-        ? InnerType extends keyof Types
-            ? Types[InnerType][]
-            : Defer<`${SchemaName}.${T}`>
-        : T extends `[${infer InnerType}]`
-        ? InnerType extends keyof Types
-            ? (Types[InnerType] | null)[]
-            : Defer<`${SchemaName}.${T}`>
-        : Defer<`${SchemaName}.${T}`>
+    // @ts-ignore
+    Carry extends string | null = (
+        RemovePrefixArrayBracket<T> extends `${infer Name}]${string}` ? Name : T
+    ) extends infer Name
+        ? Name extends `${infer A}!${string}`
+            ? A
+            : Name | null
+        : never
+    // @ts-ignore
+> = T extends `[${infer Rest}` ? CreateArray<Rest, [Carry]> : Carry
+
+type FormatType<T extends string> = FirstWord<T> extends infer T
+    ? T extends `[${string}`
+        ? // ? Is Array
+          T extends `${infer Type}!`
+            ? CreateArray<Type>
+            : CreateArray<T> | null
+        : // ? Not Array
+        T extends `${infer Type}!`
+        ? Type
+        : T | null
     : never
 
 type MapArgument<
-    SchemaName extends string,
     T extends string,
     Known extends CustomTypes = {},
     Carry extends Record<string, unknown> = {}
 > = T extends `${infer Name}:${infer Type}${'\n' | ','}${infer Rest}`
     ? MapArgument<
-          SchemaName,
           Rest,
           Known,
           Type extends `${infer _}!${infer _}`
               ? Carry & {
-                    [name in TrimLeft<Name>]: MapType<
-                        `${SchemaName}.${Name}`,
-                        FirstWord<TrimLeft<Type>>,
-                        Known
+                    [name in TrimLeft<Name>]: FormatType<
+                        FirstWord<TrimLeft<Type>>
                     >
                 }
               : Carry & {
-                    [name in TrimLeft<Name>]?: MapType<
-                        `${SchemaName}.${Name}`,
-                        FirstWord<TrimLeft<Type>>,
-                        Known
+                    [name in TrimLeft<Name>]?: FormatType<
+                        FirstWord<TrimLeft<Type>>
                     >
                 }
       >
     : T extends `${infer Name}:${infer Type}`
     ? Type extends `${infer _}!${infer _}`
         ? Carry & {
-              [name in TrimLeft<Name>]: MapType<
-                  `${SchemaName}.${Name}`,
-                  FirstWord<TrimLeft<Type>>,
-                  Known
-              >
+              [name in TrimLeft<Name>]: FormatType<FirstWord<TrimLeft<Type>>>
           }
         : Carry & {
-              [name in TrimLeft<Name>]?: MapType<
-                  `${SchemaName}.${Name}`,
-                  FirstWord<TrimLeft<Type>>,
-                  Known
-              >
+              [name in TrimLeft<Name>]?: FormatType<FirstWord<TrimLeft<Type>>>
           }
     : Carry
+
+type MapFragment<
+    Typed extends Record<string, unknown> & {
+        Fragment: Record<
+            string,
+            {
+                Target: string
+                Value: string
+            }
+        >
+    }
+> = Typed extends { Fragment: infer Fragments }
+    ? Omit<Typed, 'Fragment'> & {
+          Fragment: Prettify<{
+              [K in keyof Fragments]: Fragments[K] extends {
+                  Target: infer Target extends string
+                  Value: infer Value
+              }
+                  ? Typed extends {
+                        [a in Target]: infer Schema extends {
+                            [a in Extract<Value, string>]: unknown
+                        }
+                    }
+                      ? Prettify<Pick<Schema, Extract<Value, string>>>
+                      : {}
+                  : {
+                        K: K
+                        F: Fragments
+                    }
+          }>
+      }
+    : Typed
 
 /**
  * Infers GraphQL types to TypeScript
@@ -425,9 +380,10 @@ type MapArgument<
 export type CreateMobius<
     T extends string,
     Scalars extends Scalar = {}
-> = CreateInnerMobius<T, Scalars> extends infer Typed
+> = CreateInnerMobius<T> extends infer Typed
     ? Prettify<
-          Typed &
+          // @ts-ignore
+          MapFragment<ResolveType<Typed, Scalars>> &
               ('Query' extends keyof Typed
                   ? {}
                   : {
@@ -447,9 +403,73 @@ export type CreateMobius<
                   ? {}
                   : {
                         Fragment: {}
-                    })
+                    }) &
+              Scalars
       >
     : never
+
+// ? Quick way to display an error (interface so it displays it correctly)
+interface Err<T> {
+    [key: PropertyKey]: never
+}
+
+type MData = string | Function | null | [MData]
+
+type UnwrapKey<
+    K extends MData,
+    Result extends Record<string, unknown>,
+    Scalars extends Record<string, unknown>,
+    Nullable = null extends K ? null : never
+> = NonNullable<K> extends (infer Next extends MData)[]
+    ? UnwrapKey<Next, Result, Scalars>[] | Nullable
+    : ResolveKey<K & string, Result, Scalars> | Nullable
+
+type ResolveKey<
+    K extends string,
+    Result extends Record<string, unknown>,
+    Scalars extends Record<string, unknown>
+> = K extends (p: infer Params) => infer Returned
+    ? {
+          [K in keyof Params]: ResolveKey<
+              // @ts-ignore: Trust me bro
+              NonNullable<Params[K]>,
+              Result,
+              Scalars
+          >
+      } extends infer Argument
+        ? Partial<Argument> extends Argument
+            ? // @ts-ignore: Trust me bro
+              (p?: Argument) => UnwrapKey<Returned, Result, Scalars>
+            : // @ts-ignore: Trust me bro
+              (p: Argument) => UnwrapKey<Returned, Result, Scalars>
+        : never
+    : K extends keyof Scalars
+    ? Scalars[K]
+    : K extends keyof Result
+    ? Result[K]
+    : Err<["Couldn't resolve the key ", K]>
+
+interface ResolveInnerType<
+    Data extends Record<string, unknown | Record<string, MData>>,
+    Scalars extends Record<string, unknown> = {}
+> {
+    result: {
+        [KI in keyof Data]: Data[KI] extends Record<string, MData>
+            ? {
+                  [KJ in keyof Data[KI]]: UnwrapKey<
+                      Data[KI][KJ],
+                      this['result'],
+                      Scalars & GQLTypes
+                  >
+              }
+            : Data[KI]
+    }
+}
+
+export type ResolveType<
+    Data extends Record<string, unknown | Record<string, MData>>,
+    Scalars extends Record<string, unknown> = {}
+> = ResolveInnerType<Data, Scalars>['result']
 
 type Selective<T> = T extends object
     ? {
@@ -457,6 +477,7 @@ type Selective<T> = T extends object
       } & ('where' extends keyof T
           ? T['where'] extends NonNullable<T['where']>
               ? {
+                    // @ts-ignore: always with where
                     select: T['Select']
                     where: T['where']
                 }
@@ -474,71 +495,87 @@ type UnwrapArray<T> = T extends Array<infer R>
 /**
  * Create Prisma-like argument syntax for Client
  */
-export type CreateQuery<T extends Record<string, unknown>> = {
-    [K in keyof T]: T[K] extends (_: infer Params) => infer Query
-        ? UnwrapArray<Query> extends Record<string, unknown>
-            ? {
-                  select: CreateQuery<UnwrapArray<Query>>
-                  where: Params
-              }
-            : {
-                  select: true | undefined | null
-                  where: T[K] extends (_: infer Params) => any ? Params : never
-              }
-        : NonNullable<UnwrapArray<T[K]>> extends infer Query extends Record<
-              string,
-              unknown
-          >
-        ? {} extends UnwrapArray<Query>
-            ? true | undefined | null
-            : CreateQuery<UnwrapArray<Query>>
-        : true | undefined | null
-} & {
-    __typename?: true | undefined | null
-}
+export type CreateQuery<T extends Record<string, unknown>> =
+    (NonNullable<T> extends infer T
+        ? {
+              [K in keyof T]: T[K] extends (_: infer Params) => infer Query
+                  ? UnwrapArray<NonNullable<Query>> extends Record<
+                        string,
+                        unknown
+                    >
+                      ? UnwrapArray<Query> extends infer A extends Record<
+                            string,
+                            unknown
+                        >
+                          ? {
+                                select: CreateQuery<A>
+                                where: Params
+                            }
+                          : {}
+                      : {
+                            select: true | undefined | null
+                            where: T[K] extends (_: infer Params) => any
+                                ? Params
+                                : never
+                        }
+                  : NonNullable<
+                        UnwrapArray<T[K]>
+                    > extends infer Query extends Record<string, unknown>
+                  ? {} extends UnwrapArray<Query>
+                      ? true | undefined | null
+                      : CreateQuery<UnwrapArray<Query>>
+                  : true | undefined | null
+          }
+        : never) & {
+        __typename?: true | undefined | null
+    }
 
 type UnwrapFunctionalSchema<
     Schema extends Record<string, unknown> | Function | null
 > = Schema extends ((
     ...p: any[]
-) => infer Returned extends Record<string, unknown>)
-    ? Returned
+) => infer Returned extends Record<string, unknown> | null)
+    ? NonNullable<Returned>
     : Schema extends Record<string, unknown>
     ? Schema
     : never
 
 type Resolve<
     Query extends Record<string, unknown>,
-    Model extends Record<string, unknown>
-> = Prettify<{
-    [K in keyof Query]: Model extends Record<
-        K,
-        infer Schema extends Record<string, unknown> | Function | null
-    >
-        ? Query[K] extends true
-            ? Model[K]
-            : Query[K] extends {
-                  select: infer Selected extends Record<string, unknown>
-              }
-            ?
-                  | Resolve<Selected, UnwrapFunctionalSchema<Schema>>
-                  | (null extends Schema ? null : never)
-            : Query[K] extends Record<string, unknown>
-            ?
-                  | Resolve<Query[K], UnwrapFunctionalSchema<Schema>>
-                  | (null extends Schema ? null : never)
-            : {}
-        : K extends keyof Model
-        ? Model[K] extends Array<any>
-            ? K extends keyof Query
-                ? Resolve<
-                      Query[K] extends Record<string, unknown> ? Query[K] : {},
-                      Model[K][number]
-                  >[]
-                : []
-            : Model[K]
-        : never
-}>
+    M extends Record<string, unknown>
+> = NonNullable<M> extends infer Model
+    ? Prettify<{
+          [K in keyof Query]: Model extends Record<
+              K,
+              infer Schema extends Record<string, unknown> | Function | null
+          >
+              ? Query[K] extends true
+                  ? Model[K]
+                  : Query[K] extends {
+                        select: infer Selected extends Record<string, unknown>
+                    }
+                  ?
+                        | Resolve<Selected, UnwrapFunctionalSchema<Schema>>
+                        | (null extends Schema ? null : never)
+                  : Query[K] extends Record<string, unknown>
+                  ?
+                        | Resolve<Query[K], UnwrapFunctionalSchema<Schema>>
+                        | (null extends Schema ? null : never)
+                  : {}
+              : K extends keyof Model
+              ? Model[K] extends Array<any>
+                  ? K extends keyof Query
+                      ? Resolve<
+                            Query[K] extends Record<string, unknown>
+                                ? Query[K]
+                                : {},
+                            Model[K][number]
+                        >[]
+                      : []
+                  : Model[K]
+              : never
+      }>
+    : never
 
 /**
  * Create Prisma-like function for GraphQL
@@ -591,11 +628,12 @@ export const mobiusToGraphQL = <
             )
                 return value.select
 
-            const mapped = {}
+            const mapped: Record<string, unknown> = {}
 
             for (const [key, child] of Object.entries(value)) {
-                if (typeof child === 'object' && 'where' in child) {
+                if (typeof child === 'object' && 'where' in child!) {
                     mapped[`${key}(${JSON.stringify(child.where)})`] =
+                        // @ts-ignore select is always with where
                         child.select
                     continue
                 }
@@ -651,12 +689,13 @@ export const createFragment = (schema: string) => {
     const matches = schema.match(extractFragment)
     if (!matches) return {}
 
-    const fragments = {}
+    const fragments: Record<string, Record<string, true>> = {}
 
     if (matches) {
         for (const match of matches) {
-            const [, name, content] = extractFragment.exec(matches)
-            const current = {}
+            // @ts-ignore
+            const [, name, content] = extractFragment.exec(matches)!
+            const current: Record<string, true> = {}
 
             for (const item of content.split(/(,|\n)/g))
                 current[item.trim()] = true
@@ -668,11 +707,11 @@ export const createFragment = (schema: string) => {
     return fragments
 }
 
-type ToSelectiveFragment<T extends Record<string, unknown>> = {
+type ToSelectiveFragment<T extends Record<string, unknown>> = Prettify<{
     [K in keyof T]: T[K] extends Record<string, unknown>
         ? ToSelectiveFragment<T[K]>
         : true
-}
+}>
 
 export class Mobius<
     Declaration extends string = '',
@@ -685,7 +724,7 @@ export class Mobius<
     /**
      * ! For type declaration only
      */
-    mobius: TypeDefs | null = null
+    klein: TypeDefs | null = null
     /**
      * Available if `typeDefs` is passed
      */
@@ -698,14 +737,15 @@ export class Mobius<
             typeDefs?: Declaration
         }
     ) {
-        if (config?.typeDefs) this.fragments = createFragment(config.typeDefs)
+        if (config?.typeDefs)
+            this.fragments = createFragment(config.typeDefs) as any
     }
 
     protected get fetch() {
         return (
             this.config?.fetch ??
             ((query: string) =>
-                fetch(this.config?.url, {
+                fetch(this.config?.url ?? '::1', {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json'
@@ -719,8 +759,8 @@ export class Mobius<
     }
 
     $<
-        Query extends Selective<CreateQuery<TypeDefs['Query']>> = {},
-        Mutate extends Selective<CreateQuery<TypeDefs['Mutation']>> = {},
+        Query extends Selective<CreateQuery<TypeDefs['Query']>>,
+        Mutate extends Selective<CreateQuery<TypeDefs['Mutation']>>,
         Subscription extends Selective<CreateQuery<TypeDefs['Subscription']>>
     >(params: {
         query?: Query
@@ -748,10 +788,11 @@ export class Mobius<
         mutation: string
         subscription: string
     } {
+        // Too hard to resolve selective query to enforce, not worth
         const q = {
-            query: mobiusToGraphQL('query', params),
-            mutation: mobiusToGraphQL('mutation', params),
-            subscription: mobiusToGraphQL('subscription', params)
+            query: mobiusToGraphQL('query', params as any),
+            mutation: mobiusToGraphQL('mutation', params as any),
+            subscription: mobiusToGraphQL('subscription', params as any)
         }
 
         return {
@@ -760,7 +801,7 @@ export class Mobius<
                 query: this.fetch(q.query),
                 mutation: this.fetch(q.mutation),
                 subscription: this.fetch(q.subscription)
-            }
+            } as any
         }
     }
 
@@ -772,11 +813,12 @@ export class Mobius<
             {} extends Query ? {} : Resolve<Query, TypeDefs['Query'] & Scalars>
         > | null>
     } {
+        // @ts-ignore
         const q = mobiusToGraphQL('query', params)
 
         return {
             query: q,
-            result: this.fetch(q)
+            result: this.fetch(q) as any
         }
     }
 
@@ -792,11 +834,12 @@ export class Mobius<
             >
         >
     } {
+        // @ts-ignore
         const q = mobiusToGraphQL('mutate', params)
 
         return {
             mutate: q,
-            result: this.fetch(q)
+            result: this.fetch(q) as any
         }
     }
 
@@ -814,11 +857,12 @@ export class Mobius<
             >
         >
     } {
+        // @ts-ignore
         const q = mobiusToGraphQL('subscription', params)
 
         return {
             subscription: q,
-            result: this.fetch(q)
+            result: this.fetch(q) as any
         }
     }
 }