+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.1.0 - 5 Aug 2023
+Feature:
+- Support backward, and circular type thanks to @ggrandi04
+
+Change:
+- Remove unused dependencies
+
+Bug fix:
+- Query root array type safety
+- Query nullable type
+- Type inference not showing when referece
+- Fix TypeScript compile
+- Much more but I forgor 💀
+
 # 0.0.0
 Feat:
 - Init